Moved all specs to /specs dir and moved api routes to use sub-routes (ex: /api/check-for-alerts -> /api/check/alerts)
