+++ specs/cron/check/plugins/_PLUGIN_VERSIONS_CAPTURE-SPEC.md
