+++ src/app/admin/status/Status.Actions.ts
@@ -16,7 +16,7 @@ import {
     sendServerCommand,
 } from '@/utils/rust/rustServerCommands';
 import { parsePluginOutput } from '@/app/api/cron/check-plugins/Plugin.Parser';
-import { comparePluginVersions, compareVersions } from '@/app/api/cron/check-plugins/Plugin.Versions';
+import { comparePluginVersions, compareVersions } from '@/app/api/cron/check/plugins/Plugin.Versions';
 
 interface PluginInfo {
     name: string;
