+++ src/app/api/cron/check/alerts/route.ts
