+++ src/app/api/cron/check/plugins/route.ts
